Added Probability Calculation Units
These changes will enable the program to keep track of the location and
the probability information it contains.

Probability.h/.cc: is where the probability of pit, wumpus, gold is
stored.

LocationProbability.h/.cc: is where the probability associated with
location is stored. It is essensially a pair of location and probability

Makefile: added files Probability.h/.cc and LocationProbability.h/.cc in
there.